pyproject.toml
[project]
name = "ibm-wbm-code-compare"
version = "0.1.2"
description = "Lightweight webMethods branch comparison tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pre-commit>=3.5.0",
    "ruff>=0.1.9",
]

[project.scripts]
ibm-wbm-code-compare = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4"]
ignore = []

[tool.uv]
managed = true

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.14",
    "bump-my-version>=0.16.2",
]

[tool.bumpversion]
current_version = "0.1.2"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
tag = false
commit = false
allow_dirty = true

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "version.txt"
